_tools/aikb-search/requirements.txt
fastembed>=0.3.0
fastmcp>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
